feat: add /api/v1/alerts endpoint
